Create a dataframe showing percent proportions

Creates a new data frame with three columns, listing the unique values in a column for the input data frame, and the number of observations for each value, along with its proportion of the total data.

count_proportion(data, to_prop, descriptor)

Arguments

data

A data frame or data frame extension (e.g. a tibble).

to_prop

A string name of the column to be proportionalized.

descriptor

A string that can be added to describe from what data the proportion originated from.

Value

A data frame. The output has the following properties:

The first column (has the same name as the to_prop parameter) lists the unique values from the to_prop named column of the input data frame.

The second column (has the name of the descriptor + "_n") lists the number of observations for each value from the first column.

The third column (has the name of the descriptor + "_prop") lists the percent proportion of the column 1 value out of the whole original dataset.

Examples

one_value_2 <- data.frame(first_column = c("class1", "class1"))
count_proportion(one_value_2, 'first_column', 'one')
#> # A tibble: 1 × 3
#>   first_column one_n one_prop
#>   <chr>        <int>    <dbl>
#> 1 class1           2      100